REFACT: extract json schema compiler logic to a base class
